organize components into src/layout/ and src/ui/ directories
